+++ src/main/java/maximum/industries/models/Model004.kt
@@ -36,7 +36,7 @@ class Model004 : IModel {
                 .updater(Nesterovs(learningRate))
                 .graphBuilder()
                 .addInputs("input")
-                .setInputTypes(InputType.convolutional(sz, sz, inChannels))
+                .setInputTypes(InputType.convolutional(sz.toLong(), sz.toLong(), inChannels.toLong()))
 
                 // ######## First convolution #############################
                 .F("conv1", "input") {
