+++ src/main/java/maximum/industries/models/Model007.kt
@@ -56,7 +56,7 @@ class Model007 : IModel {
                 .updater(Nesterovs(learningRate))
                 .graphBuilder()
                 .addInputs("input")
-                .setInputTypes(InputType.convolutional(sz, sz, inChannels))
+                .setInputTypes(InputType.convolutional(sz.toLong(), sz.toLong(), inChannels.toLong()))
 
                 // ######## First convolution #############################
                 .F("conv1", "input") {
