+++ src/main/java/maximum/industries/models/Model008.kt
@@ -55,7 +55,7 @@ class Model008 : IModel {
                 .updater(Nesterovs(learningRate))
                 .graphBuilder()
                 .addInputs("input")
-                .setInputTypes(InputType.convolutional(sz, sz, inChannels))
+                .setInputTypes(InputType.convolutional(sz.toLong(), sz.toLong(), inChannels.toLong()))
 
                 // ######## First convolution #############################
                 .F("conv1", "input") {
