+++ src/main/java/maximum/industries/models/Model012.kt
@@ -37,7 +37,7 @@ class Model012 : IModel {
                 .updater(Adam(learningRate))
                 .graphBuilder()
                 .addInputs("input")
-                .setInputTypes(InputType.convolutional(sz, sz, inChannels))
+                .setInputTypes(InputType.convolutional(sz.toLong(), sz.toLong(), inChannels.toLong()))
 
                 // ######## First convolution #############################
                 .F("conv1", "input") {
